fixed the map
user can name pins and add a brief desc on click of the map. have to fill out fields first before tapping. It's ugly but itll do. maybe change the color to white idk. also saves to a pin object and a savedpins list. confused on binding pins to maps but overall works.